refactor(api): move authenticationSessionService
to identity-access-management bounded context
