+++ api/tests/acceptance/application/authentication/oidc/token-route-post_test.js
@@ -3,8 +3,8 @@ import querystring from 'node:querystring';
 import jsonwebtoken from 'jsonwebtoken';
 
 import { AuthenticationSessionContent } from '../../../../../lib/domain/models/AuthenticationSessionContent.js';
-import * as authenticationSessionService from '../../../../../lib/domain/services/authentication/authentication-session-service.js';
 import { oidcAuthenticationServiceRegistry } from '../../../../../lib/domain/usecases/index.js';
+import { authenticationSessionService } from '../../../../../src/identity-access-management/domain/services/authentication-session.service.js';
 import { createServer, databaseBuilder, expect, sinon } from '../../../../test-helper.js';
 
 const uuidPattern = new RegExp(/^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i);
